V 1.0.0 -add sort events -new pages

+++ src/main/java/ru/zulvit/databasecoursework/repository/InvitationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.zulvit.databasecoursework.model.Invitation;
 
 @Repository
-public interface InvitationRepository extends JpaRepository<Invitation, Integer> {
+public interface InvitationRepository extends JpaRepository<Invitation, Long> {
 }
